code/wsgi-middleware/django_polls_wsgi_wrap/pyproject.toml
[tool.poetry]
name = "django_polls"
version = "0.1.0"
description = ""
authors = ["Amit Saha <mail@echorand.me>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
opentelemetry-instrumentation-wsgi = "^0.30b1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
